Move NodeComparator up one package
