+++ modules/pipeline/prefix-span-pattern-detector/src/integration-test/kotlin/org.cafejojo.schaapi.pipeline.patterndetector.prefixspan/PrefixSpanJimpleNodeIntegrationTest.kt
@@ -2,7 +2,7 @@ package org.cafejojo.schaapi.pipeline.patterndetector.prefixspan
 
 import com.nhaarman.mockito_kotlin.mock
 import org.assertj.core.api.Assertions.assertThat
-import org.cafejojo.schaapi.models.libraryusagegraph.jimple.compare.GeneralizedNodeComparator
+import org.cafejojo.schaapi.models.libraryusagegraph.jimple.GeneralizedNodeComparator
 import org.jetbrains.spek.api.Spek
 import org.jetbrains.spek.api.dsl.describe
 import org.jetbrains.spek.api.dsl.it
